pyblog: Allow opening interactive shell when deploying
This almost always happens when I gendeploy: gen asks me to touch the
new post which I agree, but then the project root is dirty and I'm
greeted with the continue or not prompt, at which point I have no choice
but to open a new shell (or ^Z suspend the current job) to commit the
changes.

This commit allows to open an interactive shell in place when project
root is found to be dirty when deploying, which nicely solves the issue.

Closes #10.