pyblog: Fix bug in exclude list feature
Previously exclude_list is a list of relative paths, and it is hard to
deal with "./" resulted from joining an empty root with a name, thus
exclude list items in the root directory (e.g., ./template.html) aren't
really excluded. Not to mention the awkward trailing slash handling, and
other resolution issues (mainly ..).

In this commit we switch to an absolute exclude_list, eliminating the
aforementioned bug and simplifing exclude_list queries.